+++ packages/apollo-link/src/linkUtils.ts
@@ -1,5 +1,5 @@
 import { getOperationName } from 'apollo-utilities';
-import Observable from 'zen-observable-ts';
+import * as Observable from 'zen-observable';
 import { print } from 'graphql/language/printer';
 
 import { GraphQLRequest, Operation } from './types';
@@ -34,7 +34,7 @@ export function isTerminating(link: ApolloLink): boolean {
   return link.request.length <= 1;
 }
 
-export function makePromise<R>(observable: Observable<R>): Promise<R> {
+export function toPromise<R>(observable: Observable<R>): Promise<R> {
   let completed = false;
   return new Promise<R>((resolve, reject) => {
     observable.subscribe({
@@ -50,6 +50,20 @@ export function makePromise<R>(observable: Observable<R>): Promise<R> {
       },
       error: reject,
     });
+  });
+}
+
+// backwards compat
+export const makePromise = toPromise;
+
+export function fromPromise<T>(promise: Promise<T>): Observable<T> {
+  return new Observable<T>(observer => {
+    promise
+      .then((value: T) => {
+        observer.next(value);
+        observer.complete();
+      })
+      .catch(observer.error.bind(observer));
   });
 }