+++ packages/apollo-link/src/types.ts
@@ -1,4 +1,4 @@
-import Observable from 'zen-observable-ts';
+import * as Observable from 'zen-observable';
 import { ExecutionResult, DocumentNode } from 'graphql';
 
 export interface GraphQLRequest {