+++ packages/apollo-link-http/CHANGELOG.md
@@ -1,4 +1,11 @@
 # Change log
 
 ### vNEXT
+
+### v0.9.0
+- changed `fetcherOptions` to be `fetchOptions` and added a test for using 'GET' requests
+
+### v0.8.0
+- throw error on empty ExectionResult (missing)
+- support setting credentials, headers, and fetcherOptions in the setup of the link
 - removed sending of context to the server and allowed opt-in of sending extensions